refactor: 검증 담당 클래스의 디렉토리 변경
